backend/requirements.txt
# ─── LangChain & LangGraph ────────────────────────────────────────────────────
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.23
langchain-core==0.3.59
langgraph==0.3.34

# ─── API server ───────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.40
asyncpg==0.31.0
alembic==1.15.2

# ─── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── RAG / Embeddings ─────────────────────────────────────────────────────────
chromadb==0.6.3
sentence-transformers==3.4.1

# ─── Document parsing ─────────────────────────────────────────────────────────
pdfplumber==0.11.6
docx2txt==0.8
openpyxl==3.1.5
python-magic==0.4.27

# ─── Config & utilities ───────────────────────────────────────────────────────
pydantic==2.13.0
pydantic-settings==2.7.1
python-dotenv==1.1.0
structlog==25.3.0
httpx==0.28.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
httpx==0.28.1
